==> analysis/04-variation-drivers.Rmd <==
---
title: "Determining drivers of variation across datasets"
author: "unawaz1996"
date: "2023-09-19"
output:
  html_notebook: default
  workflowr::wflow_html:
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    autodep = TRUE,
    eval = TRUE,
	echo = TRUE,
	warning = FALSE,
	message = FALSE
)
```


# Introduction

This R markdown document contains the complete code for determing drivers of variation of bulk RNA-sequencing files for Brain Integrative Transcriptome Hub (BITHub). All datasets were preprocessed using the code in `01-bulk-preprocess-data.Rmd`, important metadata were retained in `03-bulk-preprocess-data.Rmd` and deconvolution analysis was performed in `04-bulk-deconvolution.Rmd`. 

In order to rerun the scripts - make sure to change the directories to the location of where the input files for preprocessing are stored. All directories are defined in each dataset section. 

All other files are stored within the data folder of this R workflow directory. 

```{r}
source("code/preprocess/functions.R")
library(recount3)
library(magrittr)
library(tibble)
library(reshape2)
library(SummarizedExperiment)
library(corrplot)
library(dplyr)
library(ggvenn)
library(glue)
library(pander)
library(gridExtra)
library(variancePartition)
library(DT)
library(EnsDb.Hsapiens.v86)
library(singscore)
library(ggplot2)
library(UpSetR)
library(patchwork)
library(pheatmap)
library(ggpubr)
library(variancePartition)
source("code/preprocess/raw_file_locations.R")
```

```{r}
gene2symbol = genes(EnsDb.Hsapiens.v86, return.type="DataFrame") %>%
    as.data.frame()
gene2symbol %<>% 
    dplyr::select("EnsemblID" = "gene_id", gene_name)
```

```{r load-data}
e <- list()
  
e$Raw$GTEx <- read.csv(file.path(gtex.out, "GTEx-exp.csv"), 
                       check.names = FALSE, row.names = 1)
e$Raw$PE <- read.csv(file.path(pe.out, "PsychEncode-exp.csv"), 
                     row.names =1, check.names = FALSE)
e$Raw$BSpan <- read.csv(file.path(bspan.out, "BrainSpan-exp.csv"),
                        check.names = FALSE, row.names = 1)
e$Raw$BSeq <- read.csv(file.path(bseq.out,"BrainSeq-exp.csv"), 
                       check.names = FALSE, row.names = 1)

e$Raw$HDBR <- read.csv(file.path(hdbr.out, "HDBR-exp.csv"), 
                       check.names = FALSE, row.names = 1)  %>% 
  column_to_rownames("EnsemblID")

#e$Raw$Ram <- read.csv(file.path(brain.dir, "/Ramakar/Formatted/Ramaker-exp.csv"), 
 #                      check.names = FALSE, row.names = 1)


#e$Raw <- lapply(e$Raw, function(x) {
 #   rownames(x) <- x[,1]
  #  x <- x[,-1]
   # return(x)
#  })
```

```{r metadata-load}
md = list()
md$GTEx <- read.csv(file.path(gtex.out, "GTEx-metadata.csv"), 
                       check.names = FALSE, row.names = 1) %>% 
  left_join( read.csv(file.path(gtex.out, 
                                "GTEx-decon.csv"), 
                      header=TRUE, row.names = 1) %>%
               rownames_to_column("SampleID"), by="SampleID")

  
md$PE <- read.csv(file.path(pe.out ,"PsychEncode-metadata.csv"),
                  check.names = FALSE, row.names=1) %>% 
  left_join( read.csv(file.path(pe.out, 
                                "PsychEncode-decon.csv"), 
                      header=TRUE, row.names = 1) %>%
               rownames_to_column("SampleID"), by="SampleID")

md$BSeq <- read.csv(file.path(bseq.out,"BrainSeq-metadata.csv"), 
                       check.names = FALSE) %>% 
  left_join( read.csv(file.path(bseq.out, 
                                "BrainSeq-decon.csv"), 
                      header=TRUE, row.names = 1) %>%
               rownames_to_column("SampleID"), by="SampleID")


md$HDBR <- read.csv(file.path(hdbr.out, "HDBR-metadata.csv"), 
                       check.names = FALSE, row.names = 1)  %>% 
  left_join( read.csv(file.path(hdbr.out, 
                                "HDBR-decon.csv"), 
                      header=TRUE, row.names = 1) %>%
               rownames_to_column("SampleID"), by="SampleID")


md$BSpan <- read.csv(file.path(bspan.out,
                               "BrainSpan-metadata.csv"),
                     check.names = FALSE, row.names =1) %>% 
  left_join( read.csv(file.path(bspan.out, 
                                "BrainSpan-decon.csv"), 
                      header=TRUE, row.names = 1) %>%
               rownames_to_column("SampleID"), by="SampleID")
  

```


# Determining drivers of variation 

To investigate the impact of biological and technical factors on gene expression, I performed an initial exploratory analysis on all bulk RNA-seq datasets grouped together via a PCA. The count matrices from all bulk RNA-seq datasets were filtered to remove lowly expressed genes. The expression cut-off was selected at 1 RPKM/TPM in at least 10% of the samples 


```{r}
thresh <- function(x) {
      y <- x > 1
      keep <- rowSums(y) > (ncol(y) / 10)
      return(x[keep,])
    }
```

```{r}
gtex <- thresh(e$Raw$GTEx)
span <- thresh(e$Raw$BSpan)
seq <- thresh(e$Raw$BSeq)
pe <- thresh(e$Raw$PE)
hdbr <- thresh(e$Raw$HDBR)

```

```{r}
common <- intersect(rownames(gtex), rownames(span))
common <- intersect(common, rownames(seq))
common <- intersect(common, rownames(pe))
common <- intersect(common, rownames(hdbr))
#common <- intersect(common, rownames(ram))


merge <- cbind(gtex[common,], 
               span[common,], 
               seq[common,], 
               pe[common,],
               hdbr[common,])
```

An interquantile normalization was perfromed to group the samples together. This would allow me to merge the filtered count matrices into a single-data frame and make the distributions across datasets identical. 

## Principal component analysis of all data 

```{r}
library(preprocessCore)
q.merge <- normalize.quantiles(as.matrix(merge), copy = FALSE)
  
  # log transform
q.merge <- log2(q.merge + 0.5)
```

```{r}
pca <- princomp(q.merge, cor = TRUE) # ~10 minutse
```


```{r}
m.merge = list()
m.merge$BSeq= md$BSeq %>% 
  dplyr::select("SampleID", "AgeInterval", "Sex", "Period",
                "Regions", "StructureAcronym", "DonorID")  %>% 
  mutate(Dataset = c("Brainseq"))
m.merge$BSpan = md$BSpan %>% 
  dplyr::select("SampleID", "AgeInterval", "Sex", "Period",
                "Regions", "StructureAcronym", "DonorID")  %>% 
  mutate(Dataset = c("BrainSpan"))

m.merge$GTEx = md$GTEx %>% 
  dplyr::select("SampleID", "AgeInterval", "Sex", "Period",
                "Regions", "StructureAcronym", "DonorID") %>% 
  mutate(Dataset = c("GTEx"))
m.merge$PE = md$PE %>% 
  dplyr::select("SampleID", "AgeInterval", 
                "Sex", "Period", "Regions", "StructureAcronym", "DonorID")  %>% 
  mutate(Dataset = c("PsychEncode"))

m.merge$HDBR = md$HDBR %>% 
  dplyr::select("SampleID", "AgeInterval", "Sex", 
                "Regions", "StructureAcronym", "DonorID") %>% 
  mutate(Period = c("Prenatal"), 
         Dataset = c("HBDR"))



m.merge = do.call(rbind, m.merge)
```

```{r}
plot.data <- as.data.frame(pca$loadings[,1:3])
colnames(plot.data) <- c("PC1", "PC2", "PC3")
plot.data = plot.data[rownames(plot.data) %in% m.merge$SampleID,]
```

```{r}
library(glue)
pca = plot.data %>% 
  rownames_to_column("SampleID") %>%
  left_join(m.merge, by = "SampleID") %>%
 # dplyr::filter(Dataset != "Ramaker") %>%
  ggplot(aes(x = PC1, y = PC2, color = Dataset, shape = Period, alpha = Dataset)) +
    geom_point(size =5) + 
  scale_color_manual(values =c("Brainseq" ="#F1605F",
                               "BrainSpan" = "#49215F", 
                               "GTEx" = "#7DA1D4", 
                               "HBDR" = "#E11F28", 
                               "PsychEncode" = "#FAAD79")) +
  theme_bw() + 
   labs(
    x = glue("PC1 ({percent(pca$sdev[[1]]^2 / sum(pca$sdev^2), 0.1)})"),
    y = glue("PC2 ({percent(pca$sdev[[2]]^2 / sum(pca$sdev^2), 0.1)})"),
    colour = "Dataset",
    shape = "Period"
  ) + scale_alpha_manual(values = c("Brainseq" = 0.45, 
                                    "BrainSpan" = 0.65, 
                                    "GTEx" = 0.65,
                                    "HBDR" = 1, 
                                    "PsychEncode" = 0.65)) +
  theme(axis.text = element_text(size=12)) + scale_shape_manual(values = c("Postnatal" =0, 
                                                                           "Prenatal" = 19))

pca 
```



```{r}
ggsave(pca,
       filename = here::here("output/Thesis_plots/Variance/PCA/all-data.png"),
       height = 6,
       width = 8,
       units = "in",
       dpi = 1200)
```

```{r}
plot.data  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(m.merge) %>%
    dplyr::select(
        PC1, PC2, PC3, 
        Period, 
        Regions, 
        Dataset, 
        Sex
    ) %>% 
  drop_na(Regions) %>%
   mutate(Period = as.numeric(as.factor(Period)), 
          Regions = as.numeric(as.factor(Regions)), 
          Dataset = as.numeric(as.factor(Dataset)), 
          Sex = as.numeric(as.factor(Sex))) %>%
    cor() %>% 
    corrplot(
        type = "lower", 
        diag = FALSE, 
        addCoef.col = 1)
```

## Clustering within each individual dataset

To investigate if these clustering trends observed above are maintained within individual datasets, I determined the primary drivers of variation for each dataset seperately. 

### BrainSeq 

```{r}
pc = log(seq + 0.05) %>% 
  t() %>% 
   prcomp(scale = TRUE)
```

```{r}
pca_bseq =pc$x %>% 
  as.data.frame() %>%
  dplyr::select(PC1, PC2, PC3) %>%
   rownames_to_column("SampleID") %>% 
  left_join(md$BSeq, by = "SampleID") %>%
 # pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") %>% 
  ggplot(aes(PC1, PC2)
    
  ) +
  geom_point(aes(colour = mito_ratio, shape = Period),size=5) + 
  theme_bw()+ theme(legend.position = "top") + 
  scale_color_gradient2(low = "#ffaf7b", 
                       mid = "#d76d77", 
                       high = "#3a1c71")  + 
   labs(
    x = glue("PC1 ({percent(pc$sdev[[1]]^2 / sum(pc$sdev^2), 0.1)})"),
    y = glue("PC2 ({percent(pc$sdev[[2]]^2 / sum(pc$sdev^2), 0.1)})"),
    colour = "Mitochondrial rate",
    shape = "Period",title = "BrainSeq"
  ) + theme(axis.text = element_text(size=12, colour = "black", family = "serif"), 
            axis.title = element_text(size=15, family = "serif"), 
            legend.box.background = element_rect(color = "black"),
            legend.text = element_text(family = "serif"), 
            legend.title = element_text( family = "serif"),
            plot.title = element_text(family = "serif", size =20)) 
 

pca_bseq

ggsave(pca_bseq,
       filename = here::here("output/Thesis_plots/Variance/PCA/BrainSeq/BrainSeq-PCA-mitoR.png"),
        height = 6,
       width = 8,
       units = "in",
       dpi=1200)
```

```{r}
 cor.bseq.test =pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$BSeq) %>%
    dplyr::select(
        PC1, PC2, PC3,
        mito_ratio,
        rRNA_ratio, 
        RIN, 
        TotalNReads, 
        MappingRate, 
        Period, 
        AgeInterval,
        Diagnosis,
        Sex, 
        DonorID,
        Regions,
        Neurons,
        Astrocytes, 
        Oligodendrocytes, 
        Microglia,
        Endothelia
        
    ) %>% 
  drop_na(Regions) %>%
   mutate(Period = as.numeric(as.factor(Period)), 
          Regions = as.numeric(as.factor(Regions)),
          Sex = as.numeric(as.factor(Sex)), 
          DonorID = as.numeric(as.factor(DonorID)), 
          Diagnosis = as.numeric(as.factor(Diagnosis)),
          AgeInterval = as.numeric(as.factor(AgeInterval))) %>%
    cor.mtest()

 cor.bseq =pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$BSeq) %>%
    dplyr::select(
        PC1, PC2, PC3,
        mito_ratio,
        rRNA_ratio, 
        RIN, 
        TotalNReads, 
        MappingRate, 
        Period, 
        AgeInterval,
        Diagnosis,
        Sex, 
        DonorID,
        Regions,
        Neurons,
        Astrocytes, 
        Oligodendrocytes, 
        Microglia,
        Endothelia
        
    ) %>% 
  drop_na(Regions) %>%
   mutate(Period = as.numeric(as.factor(Period)), 
          Regions = as.numeric(as.factor(Regions)),
          Sex = as.numeric(as.factor(Sex)), 
          DonorID = as.numeric(as.factor(DonorID)), 
          Diagnosis = as.numeric(as.factor(Diagnosis)),
          AgeInterval = as.numeric(as.factor(AgeInterval))) %>%
    cor()

corrplot(cor.bseq, p.mat =cor.bseq.test$p, method = 'color', diag = FALSE, type = 'upper',
         sig.level = c(0.05), pch.cex = 0.9,
         insig = 'label_sig', pch.col = 'grey20', order = 'AOE')

cor.bseq = cor.bseq[-c(1:3),]
cor.bseq = cor.bseq[,c(1:3)]
cor.bseq.test$p = cor.bseq.test$p[-c(1:3),]
cor.bseq.test$p = cor.bseq.test$p[,c(1:3)]
cor.bseq %>% corrplot(
        method = 'color',
        p.mat =cor.bseq.test$p,
         sig.level = c(0.05), pch.cex = 1.5,
         insig = 'label_sig', pch.col = 'black', cl.pos = 'n')

```

### BrainSpan

```{r}
pc = log(span + 0.05) %>%
  t() %>%
   prcomp(scale = TRUE)
```


```{r}
pca_bspan  = pc$x %>%
  as.data.frame() %>%
  dplyr::select(PC1, PC2, PC3) %>%
   rownames_to_column("SampleID") %>%
  left_join(md$BSpan, by = "SampleID") %>%
 # pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") %>%
  ggplot(aes(PC1, PC2)

  ) +
  geom_point(aes(colour = Regions, shape = Period),size=6) +
  theme_bw()+ theme(legend.position = "top") +
  scale_color_manual(values= c("Cerebellum" = "#FFD256",
                                           "Subcortex" = "#838DE0",
                                          "Cortex" = "#E85571")) +
   labs(
    x = glue("PC1 ({percent(pc$sdev[[1]]^2 / sum(pc$sdev^2), 0.1)})"),
    y = glue("PC2 ({percent(pc$sdev[[2]]^2 / sum(pc$sdev^2), 0.1)})"),
    colour = "Regions",
    shape = "Period", title = "BrainSpan"
  ) + theme(axis.text = element_text(size=12, colour = "black", family = "serif"),
            axis.title = element_text(size=15, family = "serif"),
            legend.box.background = element_rect(color = "black"),
            legend.text = element_text(family = "serif"),
            legend.title = element_text( family = "serif"),
            plot.title = element_text(family = "serif", size =20))

pca_bspan

ggsave(pca_bspan,
       filename = here::here("output/Thesis_plots/Variance/PCA/BrainSpan/BrainSpan-PCA.png"),
        height = 6,
       width = 8,
       units = "in",
       dpi=1200)
```

```{r  fig.cap="Correlation"}
 pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$BSpan) %>%
    dplyr::select(
        PC1, PC2, PC3,
        Period,
        AgeInterval,
        DonorID,
        Regions,
        Sex,
        RIN,
       PMI,
        pH,
      Dissectionscore
    ) %>%
  drop_na(RIN, pH) %>%
   mutate(Period = as.numeric(as.factor(Period)),
          Regions = as.numeric(as.factor(Regions)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          AgeInterval= as.numeric(as.factor(AgeInterval))) %>%
    cor() %>%
    corrplot(
        type = "lower",
        diag = FALSE,
        addCoef.col = 1,
        number.cex = 0.75)

```

```{r}

cor.bspan =   pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$BSpan) %>%
    dplyr::select(
        PC1, PC2, PC3,
        Period,
        PMI,
        RIN,
        pH,
        Dissectionscore,
        DonorID,
        AgeInterval,
        Sex,
        Regions,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia) %>%
  drop_na(RIN, pH) %>%
   mutate(Period = as.numeric(as.factor(Period)),
          Regions = as.numeric(as.factor(Regions)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          AgeInterval= as.numeric(as.factor(AgeInterval))) %>%
    cor()

cor.bspan.test = pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$BSpan) %>%
    dplyr::select(
        PC1, PC2, PC3,
        Period,
        PMI,
        RIN,
        pH,
        Dissectionscore,
        DonorID,
        AgeInterval,
        Sex,
        Regions,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia) %>%
  drop_na(RIN, pH) %>%
   mutate(Period = as.numeric(as.factor(Period)),
          Regions = as.numeric(as.factor(Regions)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          AgeInterval= as.numeric(as.factor(AgeInterval))) %>%
    cor.mtest()


cor.bspan = cor.bspan[-c(1:3),]
cor.bspan = cor.bspan[,c(1:3)]
cor.bspan.test$p = cor.bspan.test$p[-c(1:3),]
cor.bspan.test$p = cor.bspan.test$p[,c(1:3)]
cor.bspan %>% corrplot(
        method = 'color',
        p.mat =cor.bspan.test$p,
         sig.level = c(0.05), pch.cex = 1.5, pch.col = 'black', cl.pos = 'n',
         insig = 'label_sig')

```

### GTEx

```{r}
pc = log(gtex + 0.05) %>%
  t() %>%
   prcomp(scale = TRUE)
```

```{r}
pca_gtex = pc$x %>%
  as.data.frame() %>%
  dplyr::select(PC1, PC2, PC3) %>%
   rownames_to_column("SampleID") %>%
  left_join(md$GTEx, by = "SampleID") %>%
 # pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") %>%
  ggplot(aes(PC1, PC2)

  ) +
  geom_point(aes(colour = as.numeric(rRNA_ratio), shape = Regions, alpha = Regions),size=5) +
  theme_bw()+ theme(legend.position = "top") +
  scale_color_gradient(low = "#EECDA3",


                       high = "#ef629f")  +
  scale_shape_manual(values = c("Cerebellum" =15,
                                "Spinal Cord" = 0,
                                "Cortex" = 17,
                                "Subcortex" = 19)) +
  scale_alpha_manual(values = c("Cerebellum" =1,
                                "Spinal Cord" = 1,
                                "Cortex" = 0.5,
                                "Subcortex" = 0.5)) +
   labs(
    x = glue("PC1 ({percent(pc$sdev[[1]]^2 / sum(pc$sdev^2), 0.1)})"),
    y = glue("PC2 ({percent(pc$sdev[[2]]^2 / sum(pc$sdev^2), 0.1)})"),
    color = "rRNA rate",
    shape = "Regions", title = "GTEx"
  ) +theme(axis.text = element_text(size=12, colour = "black", family = "serif"),
            axis.title = element_text(size=15, family = "serif"),
            legend.box.background = element_rect(color = "black"),
            legend.text = element_text(family = "serif"),
            legend.title = element_text( family = "serif"),
            plot.title = element_text(family = "serif", size =20))

pca_gtex

ggsave(pca_gtex ,
       filename = here::here("output/Thesis_plots/Variance/PCA/GTEx/GTEx-PCA-rRNA.png"),
        height = 6,
       width = 8,
       units = "in",
       dpi=1200)
```


```{r}
cor.gtex = pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$GTEx) %>%
    dplyr::select(
        PC1, PC2, PC3,
        rRNA_ratio,
        ExonicRate,
        IntronicRate,
        TotalNReads,
        MappingRate,
        SequencingBatch,
        BSS_Collection_side_code,
        RIN,
        PMI,
        Sex,
        DonorID,
        HardyScale,
        AgeInterval,
        Regions,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia

    ) %>%
  drop_na(Regions) %>%
   mutate(RIN = as.numeric(as.factor(RIN)),
          TotalNReads = as.numeric(as.factor(TotalNReads)),
          MappingRate = as.numeric(as.factor(MappingRate)),
          Regions = as.numeric(as.factor(Regions)),
          AgeInterval = as.numeric(as.factor(AgeInterval)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          rRNA_ratio = as.numeric(as.factor(rRNA_ratio)),
          SequencingBatch = as.numeric(as.factor(SequencingBatch)),
          BSS_Collection_side_code =  as.numeric(as.factor(SequencingBatch))) %>%
    cor()


cor.gtex.test = pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$GTEx) %>%
    dplyr::select(
        PC1, PC2, PC3,
        rRNA_ratio,
        ExonicRate,
        IntronicRate,
        TotalNReads,
        MappingRate,
        SequencingBatch,
        BSS_Collection_side_code,
        RIN,
        PMI,
        Sex,
        DonorID,
        HardyScale,
        AgeInterval,
        Regions,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia

    ) %>%
  drop_na(Regions) %>%
   mutate(RIN = as.numeric(as.factor(RIN)),
          TotalNReads = as.numeric(as.factor(TotalNReads)),
          MappingRate = as.numeric(as.factor(MappingRate)),
          Regions = as.numeric(as.factor(Regions)),
          AgeInterval = as.numeric(as.factor(AgeInterval)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          rRNA_ratio = as.numeric(as.factor(rRNA_ratio)),
          SequencingBatch = as.numeric(as.factor(SequencingBatch)),
          BSS_Collection_side_code =  as.numeric(as.factor(SequencingBatch))) %>%
    cor.mtest()

cor.gtex = cor.gtex[-c(1:3),]
cor.gtex = cor.gtex[,c(1:3)]
cor.gtex.test$p = cor.gtex.test$p[-c(1:3),]
cor.gtex.test$p = cor.gtex.test$p[,c(1:3)]
cor.gtex %>% corrplot(
        method = 'color',
        p.mat =cor.gtex.test$p,
         sig.level = c(0.05), pch.cex = 1.5,
         insig = 'label_sig', pch.col = 'black', cl.pos = 'n')
```

### HDBR 

```{r}
pc = log(hdbr + 0.05) %>%
  t() %>%
   prcomp(scale = TRUE)

pca_hdbr = pc$x %>%
  as.data.frame() %>%
  dplyr::select(PC1, PC2, PC3) %>%
   rownames_to_column("SampleID") %>%
  left_join(md$HDBR, by = "SampleID") %>%
 # pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") %>%
  ggplot(aes(PC1, PC2)) +
  geom_point(aes(colour = IntronicRate, shape = Regions),size=6, alpha = 0.85) +
  #geom_point(aes(colour = ContributingStudy, shape = Period),size=5) +
  theme_bw()+ theme(legend.position = "top")  +
  scale_color_gradient2(low = "#ffaf7b",
                       mid = "#d76d77",
                       high = "#3a1c71") +
  scale_shape_manual(values = c("Brain" = 16,
                                "Forebrain" = 17,
                                "Chroid plexus" = 2,
                                "Hindbrain" = 15,
                                "Midbrain" = 0,
                                "Spinal Cord" = 6)) +
   labs(
    x = glue("PC1 ({percent(pc$sdev[[1]]^2 / sum(pc$sdev^2), 0.1)})"),
    y = glue("PC2 ({percent(pc$sdev[[2]]^2 / sum(pc$sdev^2), 0.1)})"),
    color = "Intronic Rate",
    shape = "Regions", title = "HDBR"
  ) + theme(axis.text = element_text(size=12, colour = "black", family = "serif"),
            axis.title = element_text(size=15, family = "serif"),
            legend.box.background = element_rect(color = "black"),
            legend.text = element_text(family = "serif"),
            legend.title = element_text( family = "serif"),
            plot.title = element_text(family = "serif", size =20))

pca_hdbr
ggsave(pca_hdbr,
       filename = here::here("output/Thesis_plots/Variance/PCA/HDBR/HDBR-PCA-intronicRate.png"),
        height = 6,
       width = 8,
       units = "in",
       dpi=1200)

```


```{r}
pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$HDBR) %>%
    dplyr::select(
        PC1, PC2, PC3,
        IntronicRate,
        MappingRate,
        AgeInterval,
        Sex,
        ExonicRate,
        DonorID,
        TotalNReads,
        Regions,
        Hemisphere,
     #   PMI,
        SequencingBatch) %>%
  #drop_na(PMI, Regions) %>%
   mutate(#RIN = as.numeric(as.factor(RIN)),
          TotalNReads = as.numeric(as.factor(TotalNReads)),
          MappingRate = as.numeric(as.factor(MappingRate)),
          Regions = as.numeric(as.factor(Regions)),
          AgeInterval = as.numeric(as.factor(AgeInterval)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          Hemisphere = as.numeric(as.factor(Hemisphere)),
          SequencingBatch = as.numeric(as.factor(SequencingBatch))) %>%
    cor() %>%
    corrplot(
        type = "lower",
        diag = FALSE,
        addCoef.col = 1,
        number.cex = 0.75)



```


```{r}
cor.hdbr =pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$HDBR) %>%
    dplyr::select(
        PC1, PC2, PC3,
        IntronicRate,
        MappingRate,
        ExonicRate,
        TotalNReads,
        SequencingBatch,
        PMI,
        Sex,
        DonorID,
        AgeInterval,
        Regions,
        Hemisphere,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia) %>%
  drop_na(PMI, Regions) %>%
   mutate(#RIN = as.numeric(as.factor(RIN)),
          TotalNReads = as.numeric(as.factor(TotalNReads)),
          MappingRate = as.numeric(as.factor(MappingRate)),
          Regions = as.numeric(as.factor(Regions)),
          AgeInterval = as.numeric(as.factor(AgeInterval)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          Hemisphere = as.numeric(as.factor(Hemisphere)),
          SequencingBatch = as.numeric(as.factor(SequencingBatch))) %>%
    cor()

cor.hdbr.test =pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$HDBR) %>%
    dplyr::select(
        PC1, PC2, PC3,
        IntronicRate,
        MappingRate,
        ExonicRate,
        TotalNReads,
        SequencingBatch,
        PMI,
        Sex,
        DonorID,
        AgeInterval,
        Regions,
        Hemisphere,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia) %>%
  drop_na(PMI, Regions) %>%
   mutate(#RIN = as.numeric(as.factor(RIN)),
          TotalNReads = as.numeric(as.factor(TotalNReads)),
          MappingRate = as.numeric(as.factor(MappingRate)),
          Regions = as.numeric(as.factor(Regions)),
          AgeInterval = as.numeric(as.factor(AgeInterval)),
          Sex = as.numeric(as.factor(Sex)),
          DonorID = as.numeric(as.factor(DonorID)),
          Hemisphere = as.numeric(as.factor(Hemisphere)),
          SequencingBatch = as.numeric(as.factor(SequencingBatch))) %>%
    cor.mtest()
cor.hdbr = cor.hdbr[-c(1:3),]
cor.hdbr = cor.hdbr[,c(1:3)]
cor.hdbr.test$p = cor.hdbr.test$p[-c(1:3),]
cor.hdbr.test$p = cor.hdbr.test$p[,c(1:3)]
cor.hdbr %>% corrplot(
        method = 'color',
        p.mat =cor.hdbr.test$p,
         sig.level = c(0.05), pch.cex = 1.5,pch.col = 'black', cl.pos = 'n',insig = 'label_sig')
```


### PsychENCODE 


```{r}
pc = log(pe + 0.05) %>%
  t() %>%
   prcomp(scale = TRUE)



pca_pe  = pc$x %>%
  as.data.frame() %>%
  dplyr::select(PC1, PC2, PC3) %>%
   rownames_to_column("SampleID") %>%
  left_join(md$PE, by = "SampleID") %>%
 # pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") %>%
  ggplot(aes(PC1, PC2)

  ) +
  geom_point(aes(colour = ContributingStudy, shape = Period),size=6, alpha = 0.75) +
  theme_bw()+ theme(legend.position = "top") +
   labs(
    x = glue("PC1 ({percent(pc$sdev[[1]]^2 / sum(pc$sdev^2), 0.1)})"),
    y = glue("PC2 ({percent(pc$sdev[[2]]^2 / sum(pc$sdev^2), 0.1)})"),
    color = "Contributing study", title = "PsychENCODE"
  ) + theme(axis.text = element_text(size=12, colour = "black", family = "serif"),
            axis.title = element_text(size=15, family = "serif"),
            legend.box.background = element_rect(color = "black"),
            legend.text = element_text(family = "serif"),
            legend.title = element_text( family = "serif"),
            plot.title = element_text(family = "serif", size =20))

pca_pe

ggsave(pca_pe,
       filename = here::here("output/Thesis_plots/Variance/PCA/PsychEncode/PE-PCA.png"),
        height = 6,
       width = 8,
       units = "in",
       dpi = 1200)
```


```{r}
cor.pe =pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$PE) %>%
    dplyr::select(
        PC1, PC2, PC3,
        individualIDSource,
        Sex,
        ContributingStudy,
        Period,
        #brainWeight,
        AgeNumeric,
        Diagnosis,
        causeDeath,
        AgeInterval,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia
    ) %>%
  #drop_na(brainWeight) %>%
   mutate(#RIN = as.numeric(as.factor(RIN)),
          AgeInterval = as.numeric(as.factor(AgeInterval)),
          Sex = as.numeric(as.factor(Sex)),
          individualIDSource = as.numeric(as.factor(individualIDSource)),
          ContributingStudy = as.numeric(as.factor(ContributingStudy)),
          Period = as.numeric(as.factor(Period)),
          Diagnosis = as.numeric(as.factor(Diagnosis)),
          causeDeath = as.numeric(as.factor(causeDeath))) %>%
    cor()

cor.pe.test =pc$x  %>%
    as.data.frame() %>%
  rownames_to_column("SampleID") %>%
  left_join(md$PE) %>%
    dplyr::select(
        PC1, PC2, PC3,
        individualIDSource,
        Sex,
        ContributingStudy,
        Period,
        #brainWeight,
        AgeNumeric,
        Diagnosis,
        causeDeath,
        AgeInterval,
        Neurons,
        Astrocytes,
        Oligodendrocytes,
        Microglia,
        Endothelia
    ) %>%
  #drop_na(brainWeight) %>%
   mutate(#RIN = as.numeric(as.factor(RIN)),
          AgeInterval = as.numeric(as.factor(AgeInterval)),
          Sex = as.numeric(as.factor(Sex)),
          individualIDSource = as.numeric(as.factor(individualIDSource)),
          ContributingStudy = as.numeric(as.factor(ContributingStudy)),
          Period = as.numeric(as.factor(Period)),
          Diagnosis = as.numeric(as.factor(Diagnosis)),
          causeDeath = as.numeric(as.factor(causeDeath))) %>%
    cor.mtest()

cor.pe = cor.pe[-c(1:3),]
cor.pe = cor.pe[,c(1:3)]
cor.pe.test$p = cor.pe.test$p[-c(1:3),]
cor.pe.test$p = cor.pe.test$p[,c(1:3)]
cor.pe %>% corrplot(
        method = 'color',
        p.mat =cor.pe.test$p,
         sig.level = c(0.05), pch.cex = 1.5,
          pch.col = 'black', cl.pos = 'n', insig = 'label_sig')



```


## Variance partitioning analysis 

variancePartition is a statistical and visualization framework which fits a linear mixed model for each gene and partitions the total variance into the contribution of each variable in the experimental design, plus the residual variance. The fractions sum to 1 for each gene, the variance fractions are easily interpretable across genes and sources of variation. 

Prior to performing a variancePartition analysis, the correlation of all variables considered for varianceParition analysis was performed using a Canonical Correlation Analysis (CCA), where variables with a highly correlating score (correlation > 0.95) were dropped, as high correlation can overestimate the contribution of variation 

### BrainSeq

```{r}
form.bseq <- ~ AgeNumeric + (1|Regions) + (1|Sex) + RIN +  (1|Diagnosis) + mito_ratio + rRNA_ratio + TotalNReads + MappingRate +  (1|Period) + Neurons + Astrocytes + Oligodendrocytes + Endothelia + Microglia
C.bseq = canCorPairs(form.bseq, md$BSeq)
plotCorrMatrix( C.bseq )
```

### BrainSpan


```{r}
form.bspan <- ~ AgeNumeric + (1|Sex) + (1|Period) + (1|Regions) + RIN +  pH + PMI + DonorID + Dissectionscore  +
  Neurons + Astrocytes + Oligodendrocytes  + Endothelia + Microglia
C.bspan = canCorPairs(form.bspan, md$BSpan)
plotCorrMatrix( C.bspan )
```


### GTEx


```{r}
gtex.form <- ~ TotalNReads + rRNA_ratio + (1|TypeofBatch)  + (1|AgeInterval) + (1|Sex) + (1|Regions) + RIN + HardyScale + PMI + MappingRate + PMI + ExonicRate + IntronicRate + Neurons + Astrocytes + Oligodendrocytes + Microglia + Endothelia
C.gtex = canCorPairs(gtex.form, md$GTEx)
plotCorrMatrix(C.gtex)
```

```{r}
# gtex.form <- ~ TotalNReads + rRNA_ratio + (1|TypeofBatch)  + (1|AgeInterval) + (1|Sex) + (1|Regions) + RIN + HardyScale + PMI + MappingRate + PMI + ExonicRate +  Neurons + Astrocytes + Oligodendrocytes +  Endothelia
``` 

### HDBR 

```{r}
form.hdbr <- ~(1|Sex) + (1|Hemisphere) + (1|SequencingBatch)+  TotalNReads + MappingRate + (1|Regions) + PMI + DonorID + ExonicRate + IntronicRate + (1|AgeInterval) +  Neurons + Astrocytes + Oligodendrocytes + Endothelia + Microglia
C.hdbr = canCorPairs(form.hdbr, md$HDBR)
plotCorrMatrix(C.hdbr)
```

### PsychENCODE 


<!-- # ```{r} -->
<!-- # e$Scale <- lapply(e$Raw, scale) # basic z-score columnwise transformation -->
<!-- # e$Log <- lapply(e$Raw, function(x) { log2(x+0.5) }) # log2 -->
<!-- # e$Log_Scale <- lapply(e$Log, scale) # log2 followed by scale -->
<!-- # ``` -->

<!-- # ```{r} -->
<!-- # e$Thresh <- lapply(e$Raw, apply.threshold) # this is > 1 unit (TPM in GTEx, RPKM in others) in > 10% of samples -->
<!-- # e$Thresh_Scale <- lapply(e$Thresh, scale) -->
<!-- # e$Thresh_Log <- lapply(e$Thresh, function(x) { log2(x+0.5) }) -->
<!-- # e$Thresh_Log_Scale <- lapply(e$Log, scale) -->
<!-- # ``` -->

<!-- ```{r} -->
<!-- ## Histogram of median expression -->
<!--   # collect median expression values across samples for every dataset -->
<!-- #p <- lapply(e, function(x) { -->
<!-- #    # lapply(x, rowMeans) -->
<!--  #   lapply(x, function(y) { -->
<!--  #     apply(y, 1, median) -->
<!-- #    }) -->
<!-- #  }) -->

<!-- p <- lapply(e, function(x) { -->
<!--     # lapply(x, rowMeans) -->
<!--     lapply(x, function(y) { -->
<!--       apply(y, 1, mean) -->
<!--     }) -->
<!--   }) -->

<!-- p <- melt(p) -->
<!-- colnames(p) <- c("MeanExp", "Dataset", "Transform") -->
<!-- p$Dataset <- factor(p$Dataset) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # move thresholding category to a separate column -->
<!-- p$Thresh <- FALSE; -->
<!-- p$Thresh[grep("Thresh", p$Transform)] <- TRUE -->
<!-- p$Transform <- gsub("Thresh_", "", p$Transform) -->
<!-- p$Transform[which(p$Transform == "Thresh")] <- "Raw" -->
<!-- p$Transform <- factor(p$Transform, levels = c("Raw", "Log", "Scale", "Log_Scale")) -->

<!--   # ...and to separate dataframes -->
<!-- t <- p[which(p$Thresh),] -->
<!-- x <- table(t$Dataset) / 4 # this gets the number of genes in each dataset (divisor is 4 as each gene is represented by 4 transformations) -->
<!-- levels(t$Dataset) <- paste0(names(x), " (n=", x, ")") -->

<!-- p2 <- p[-which(p$Thresh),] -->
<!-- x <- table(p$Dataset) / 4 # this gets the number of genes in each dataset (divisor is 4 as each gene is represented by 4 transformations) -->
<!-- levels(p$Dataset) <- paste0(names(x), " (n=", x, ")") -->
<!-- ``` -->

<!-- ```{r} -->
<!-- no_filt =p %>%  -->
<!--   dplyr::filter(Transform == "Log" | Transform ==  -->
<!--                   "Log_Scale") %>% -->
<!--   ggplot(aes(x = MeanExp, colour = Dataset)) + -->
<!--    scale_color_manual(values = c("#F75E5E",  -->
<!--                                  "#49165E", -->
<!--                                 "#78A2EB",  -->
<!--                                "#E11F28",  -->
<!--                                  "#F9AD79"),  -->
<!--                       labels =c("BrainSeq",  -->
<!--                                 "BrainSpan",  -->
<!--                                 "PsychENCODE",  -->
<!--                                 "GTEx",  -->
<!--                                 "HDBR")) + -->
<!--     geom_density() + -->
<!--     # geom_histogram() + -->
<!--     facet_wrap(~Transform, scales = "free") + -->
<!--     theme_bw() + -->
<!--     labs(y = "Density", x = "Mean Expression Across Samples")  +  -->
<!--     theme(axis.text.x = element_text(size = 13, family = "serif", color = "black", angle = 90), -->
<!--         axis.text.y = element_text(size = 13, family = "serif", color = "black"),  -->
<!--         axis.title = element_text(size=15, family = "serif"),  -->
<!--          legend.box.background = element_rect(color = "black"), -->
<!--           legend.text = element_text(family = "serif"),  -->
<!--           legend.title = element_text( family = "serif"), -->
<!--            plot.title = element_text(family = "serif", size =20),          -->
<!--         panel.grid.major.x = element_blank(), -->
<!--         panel.grid.major.y = element_line( size=.1 ), -->
<!--               legend.position = "top")   -->

<!-- no_filt -->

<!-- #ggsave(filename = here::here("output/Thesis_plots/Attributes/exp.svg"),  -->
<!--  #      height = 6, width =12, units = "in", plot = no_filt) -->

<!-- ``` -->

<!-- # ```{r} -->
<!-- # filt =ggplot(t, aes(x = MeanExp, colour = Dataset)) + -->
<!-- #   scale_color_manual(values = c("#F75E5E",  -->
<!-- #                                  "#49165E", -->
<!-- #                                 "#78A2EB",  -->
<!-- #                                "#E11F28",  -->
<!-- #                                  "#F9AD79",  -->
<!-- #                                  "#74C69D")) + -->
<!-- #     geom_density() + -->
<!-- #     # geom_histogram() + -->
<!-- #     facet_wrap(~Transform, scales = "free") + -->
<!-- #     theme_bw() + -->
<!-- #     labs(y = "Density", x = "Median Expression Across Samples", title = ">1 RPKM in 10% of Samples") -->
<!-- #  -->
<!-- # filt -->
<!-- # ``` -->

<!-- # ```{r} -->
<!-- # ## Correspondence -->
<!-- #   p <- lapply(e[c(1,4,5,8)], function(x) { -->
<!-- #     lapply(x, function(y) { -->
<!-- #       apply(y, 1, median) -->
<!-- #     }) -->
<!-- #   }) -->
<!-- #    -->
<!-- # q <- list() -->
<!-- # q$Thresh <- q$Raw <- list() -->
<!-- # for (j in names(e$Raw)) { -->
<!-- #     q$Raw[[j]] <- cbind(p$Raw[[j]], p$Log_Scale[[j]]) -->
<!-- #     q$Thresh[[j]] <- cbind(p$Thresh[[j]], p$Thresh_Log_Scale[[j]]) -->
<!-- #   } -->
<!-- #    -->
<!-- # scatterplot <- function(x, y, lim = 100) { # x is Raw/Thresh, y is dataset -->
<!-- #     z <- q[[x]][[y]] -->
<!-- #      -->
<!-- #     # filter to set limit -->
<!-- #     rem <- (which(z[,1] > lim)) -->
<!-- #     z <- z[-rem,] -->
<!-- #      -->
<!-- #     # calculate correspondences -->
<!-- #     at1 <- round(z[which.min(abs(z[,1] - 1)), 2],2) -->
<!-- #     at10 <- round(z[which.min(abs(z[,1] - 10)), 2],2) -->
<!-- #     at25 <- round(z[which.min(abs(z[,1] - 25)), 2],2) -->
<!-- #     at100 <- round(z[which.min(abs(z[,1] - 100)), 2],2) -->
<!-- #     at <- paste(at1, at10, at25, at100, sep = "/") -->
<!-- #      -->
<!-- #     qplot(z[,1], z[,2]) + -->
<!-- #       theme_bw() + -->
<!-- #       scale_x_continuous(limits = c(NA, lim)) + -->
<!-- #       # theme(axis.title.x = element_blank(), axis.ti) -->
<!-- #       labs(x = paste0("Untransformed RPKM/TPM", "\nx=1/10/25/100 | y=", at),  -->
<!-- #            y = "Scaled log2+0.5",  -->
<!-- #            title = paste(x, y, "(n=", nrow(z), ",", length(rem), "not shown)"))  -->
<!-- #   } -->
<!-- #    -->
<!-- # p <- list() -->
<!-- #  -->
<!-- # ``` -->


<!-- # ````{r} -->
<!-- # p <- list() -->
<!-- # p$AR <- scatterplot("Raw", "PE") -->
<!-- # p$AT <- scatterplot("Thresh", "PE") -->
<!-- # p$BR <- scatterplot("Raw", "BSpan") -->
<!-- # p$BT <- scatterplot("Thresh", "BSpan") -->
<!-- # p$CR <- scatterplot("Raw", "BSeq") -->
<!-- # p$CT <- scatterplot("Thresh", "BSeq") -->
<!-- # p$DR <- scatterplot("Raw", "GTEx") -->
<!-- # p$DT <- scatterplot("Thresh", "GTEx") -->
<!-- # #p$ER <- scatter_plot("Raw", "Ram") -->
<!-- # #p$EL <- scatter_plot("Thresh", "Ram") -->
<!-- # #p$FR <- scatter_plot("Raw", "HDBR") -->
<!-- # #p$FL <- scatter_plot("Thresh", "HDBR") -->
<!-- # ``` -->




<!-- # ```{r} -->
<!-- # form.bseq <- ~ AgeNumeric + (1|Regions) + (1|Sex) + RIN +  (1|Diagnosis) + mito_ratio + rRNA_ratio + TotalNReads + MappingRate +  (1|Period) + Neurons + Astrocytes + Oligodendrocytes + Endothelia -->
<!-- #  -->
<!-- # ``` -->
<!-- #  -->
<!-- # ```{r} -->
<!-- # rownames(md$BSeq) = md$BSeq$SampleID -->
<!-- # varPar.bseq <- fitExtractVarPartModel(seq, form.bseq, md$BSeq) -->
<!-- # ``` -->
<!-- #  -->
<!-- # ```{r} -->
<!-- # varPar.bseq  = sortCols(varPar.bseq ) -->
<!-- # plotVarPart(varPar.bseq ) -->
<!-- # write.csv(varPar.bseq , file="/home/neuro/Documents/BrainData/Bulk/Brainseq/Formatted/BrainSeq-varPart-all.csv") -->
<!-- # ``` -->



<!-- The correlation plot reveals that many of the selected covariates do not correlate highly within their respective category, and therefore we will feed these into the mixed linear model.  -->



<!-- <!-- Variance partition  --> -->


<!-- <!-- ```{r} --> -->
<!-- <!-- varPar.bspan <- fitExtractVarPartModel(span,form.bspan ,md$BSpan) --> -->
<!-- <!-- varPar.bspan = sortCols(varPar.bspan) --> -->
<!-- <!-- plot_varPart = plotVarPart(varPar.bspan) --> -->
<!-- <!-- ``` --> -->

<!-- <!-- # ```{r} --> -->
<!-- <!-- # write.csv(varPar.bspan, file="/home/neuro/Documents/BrainData/Bulk/BrainSpan/Formatted/BrainSpan-varPart-all.csv") --> -->
<!-- <!-- # ``` --> -->







<!-- <!-- # varPar.gtex <- fitExtractVarPartModel(gtex,gtex.form ,md$GTEx) --> -->
<!-- <!-- # ``` --> -->
<!-- <!-- #  --> -->
<!-- <!-- # ```{r} --> -->
<!-- <!-- # varPar.gtex = sortCols(varPar.gtex) --> -->
<!-- <!-- # plot_varPart = plotVarPart(varPar.gtex) --> -->
<!-- <!-- # ``` --> -->
<!-- <!-- #  --> -->
<!-- <!-- # ```{r} --> -->
<!-- <!-- # write.csv(varPar.gtex,"/home/neuro/Documents/BrainData/Bulk/GTEx/Formatted/GTEx-varPart-all.csv") --> -->
<!-- <!-- # ``` --> -->


<!-- - PyschEncode -->

<!-- ```{r fig.height =5, fig.width=5, fig.cap="*Assessing correlation between covariates of interest from the PsychEncode data*"} -->
<!-- pe.form <- ~ (1|Sex) + (1|Diagnosis) + AgeNumeric + (1|Ethnicity) +  -->
<!--   (1|Period) + (1|individualIDSource) + (1|Death) +  Neurons + Astrocytes + Oligodendrocytes + Endothelia -->

<!-- C.pe = canCorPairs(pe.form, md$PE) -->
<!-- plotCorrMatrix(C.pe) -->
<!-- ``` -->


<!-- <!-- ```{r} --> -->
<!-- <!-- md$PE= md$PE[!duplicated(md$PE[,c('SampleID')]),] --> -->
<!-- <!-- varPar.pe <- fitExtractVarPartModel(pe, pe.form, md$PE) --> -->


<!-- <!-- ``` --> -->


<!-- <!-- ```{r} --> -->
<!-- <!-- varPar.pe = sortCols(varPar.pe) --> -->
<!-- <!-- plot_varPart = plotVarPart(varPar.pe) --> -->
<!-- <!-- ``` --> -->

<!-- <!-- ```{r} --> -->
<!-- <!-- write.csv(varPar.pe,"/home/neuro/Documents/BrainData/Bulk/PsychEncode/Formatted/PsychEncode-varPart-all.csv") --> -->
<!-- <!-- ``` --> -->

<!-- - HDBR -->




<!-- ```{r} -->
<!-- form.hdbr <- ~(1|Sex) + (1|Hemisphere) + (1|SequencingBatch)+  TotalNReads + MappingRate + (1|Regions) + PMI + DonorID + ExonicRate + IntronicRate + (1|AgeInterval) +  Neurons + Astrocytes + Oligodendrocytes + Endothelia -->
<!-- C.hdbr = canCorPairs(form.hdbr, md$HDBR) -->
<!-- plotCorrMatrix(C.hdbr) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- varPar.hdbr <- fitExtractVarPartModel(hdbr,form.hdbr ,md$HDBR) -->
<!-- varPar.hdbr = sortCols(varPar.hdbr) -->

<!-- hdbr.varpart =plot_varPart = plotVarPart(varPar.hdbr) -->

<!-- write.csv(varPar.hdbr , file = "/home/neuro/Documents/BrainData/Bulk/HDBR/Formatted/HDBR-varPart-all.csv") -->
<!-- ``` -->


<!-- ```{r} -->
<!-- pdf(file = here::here("output/Thesis_plots/Variance/PCA/alll-pca-plots.pdf"), -->
<!--     height =20, width = 15) -->
<!-- grid.arrange(pca_bseq, pca_bspan, -->
<!--              pca_gtex, pca_hdbr, -->
<!--             pca_pe, ncol=2, nrow=3) -->

<!-- dev.off() -->
<!-- ``` -->

<!-- ```{r} -->
<!-- ggarrange(plotCorrMatrix(C.hdbr), plotCorrMatrix(C.gtex)) -->
<!-- ``` -->


<!-- ## Determining the utlity of factors of unwanted variation -->

<!-- ```{r} -->
<!-- var_part = list() -->

<!-- var_part$BSpan = read.csv(file.path(brain.dir,"/BrainSpan/Formatted/BrainSpan-varPart-all.csv"), header=TRUE, -->
<!--                                     row.names =1)  %>% -->
<!--     as.data.frame() %>% -->
<!--     rownames_to_column("EnsemblID") %>% -->
<!--     left_join(gene2symbol, by = "EnsemblID") -->
<!-- var_part$BSeq = read.csv(file.path(brain.dir,"/Brainseq/Formatted/BrainSeq-varPart-all.csv"), -->
<!--                                    header=TRUE, row.names=1) %>% -->
<!--     as.data.frame() %>% -->
<!--     rownames_to_column("EnsemblID") %>% -->
<!--     left_join(gene2symbol, by = "EnsemblID") -->
<!-- var_part$GTEx = read.csv(file.path(brain.dir,"/GTEx/Formatted/GTEx-varPart-all.csv"), header=TRUE, -->
<!--                                    row.names=1) %>% -->
<!--     as.data.frame() %>% -->
<!--     rownames_to_column("EnsemblID") %>% -->
<!--     left_join(gene2symbol, by = "EnsemblID") -->
<!-- var_part$HDBR = read.csv(file.path(brain.dir,"/HDBR/Formatted/HDBR-varPart-all.csv"), header=TRUE, -->
<!--                                    row.names=1) %>% -->
<!--     as.data.frame() %>% -->
<!--     rownames_to_column("EnsemblID") %>% -->
<!--     left_join(gene2symbol, by = "EnsemblID") -->
<!-- var_part$PE = read.csv(file.path(brain.dir,"/PsychEncode/Formatted/PsychEncode-varPart-all.csv"), header=TRUE, -->
<!--                                    row.names=1) %>% -->
<!--     as.data.frame() %>% -->
<!--     rownames_to_column("EnsemblID") %>% -->
<!--     left_join(gene2symbol, by = "EnsemblID") -->
<!-- ``` -->

<!-- ```{r} -->
<!-- bspan =var_part$BSpan %>% dplyr::select(-gene_name) %>% -->
<!--   column_to_rownames("EnsemblID") %>% -->
<!-- plotVarPart() + ggtitle("BrainSpan") + ylab("Variance Explained (%)") + -->
<!--   theme(axis.text = element_text(size=12, family = "serif"), -->
<!--         axis.title = element_text(size=12,family = "serif"), -->
<!--         plot.title = element_text(hjust = 0, size = 15, family = "serif")) -->

<!-- bseq =var_part$BSeq %>% dplyr::select(-gene_name) %>% -->
<!--   column_to_rownames("EnsemblID") %>% -->
<!-- plotVarPart() + ggtitle("BrainSeq") + ylab("Variance Explained (%)") + -->
<!--   theme(axis.text = element_text(size=12, family = "serif"), -->
<!--         axis.title = element_text(size=12,family = "serif"), -->
<!--         plot.title = element_text(hjust = 0, size = 15, family = "serif")) -->


<!-- gtex  =var_part$GTEx %>% dplyr::select(-gene_name) %>% -->
<!--   column_to_rownames("EnsemblID") %>% -->
<!-- plotVarPart() + ggtitle("GTEx") + ylab("Variance Explained (%)") + -->
<!--   theme(axis.text = element_text(size=12, family = "serif"), -->
<!--         axis.title = element_text(size=12,family = "serif"), -->
<!--         plot.title = element_text(hjust = 0, size = 15, family = "serif")) -->


<!-- hdbr  =var_part$HDBR %>% dplyr::select(-gene_name) %>% -->
<!--   column_to_rownames("EnsemblID") %>% -->
<!-- plotVarPart() + ggtitle("HDBR") + ylab("Variance Explained (%)") + -->
<!--   theme(axis.text = element_text(size=12, family = "serif"), -->
<!--         axis.title = element_text(size=12,family = "serif"), -->
<!--         plot.title = element_text(hjust = 0, size = 15, family = "serif")) -->

<!-- pe  =var_part$PE %>% dplyr::select(-gene_name) %>% -->
<!--   column_to_rownames("EnsemblID") %>% -->
<!-- plotVarPart() + ggtitle("PsychENCODE") + ylab("Variance Explained (%)") + -->
<!--   theme(axis.text = element_text(size=12, family = "serif"), -->
<!--         axis.title = element_text(size=12,family = "serif"), -->
<!--         plot.title = element_text(hjust = 0, size = 15, family = "serif")) -->

<!-- pdf(file = here::here("output/Thesis_plots/Variance/PCA/varpart-all-update2.pdf"), height =20, width = 15) -->
<!-- grid.arrange(bseq, bspan, -->
<!--              gtex, hdbr, -->
<!--              pe, ncol=2, nrow=3) -->

<!-- dev.off() -->

<!-- ``` -->

<!-- ```{r} -->
<!-- median(var_part$BSeq$Sex) * 100 -->

<!-- ``` -->



<!-- # ```{r} -->
<!-- # upf1_1 = var_part$BSpan %>%  -->
<!-- #   dplyr::filter(gene_name == "UPF1") %>% -->
<!-- #   dplyr::select(-gene_name) %>%  -->
<!-- #   column_to_rownames("EnsemblID") %>%  -->
<!-- #   melt()  %>%  -->
<!-- #   ggplot(aes(x=variable, y = value)) + geom_bar(stat = "identity", fill = "#F16160") + -->
<!-- #   coord_flip() + theme_bw() + ylab("Fraction Variance Explained") + xlab("") + ggtitle("UPF1 - BrainSpan") + -->
<!-- #      theme(axis.text.x = element_text(size = 12, family = "serif", color = "black", angle = 90), -->
<!-- #         axis.text.y = element_text(size = 15, family = "serif", color = "black"),  -->
<!-- #         axis.title = element_text(size=15, family = "serif"),  -->
<!-- #          legend.box.background = element_rect(color = "black"), -->
<!-- #           legend.text = element_text(family = "serif"),  -->
<!-- #           legend.title = element_text( family = "serif"), -->
<!-- #            plot.title = element_text(family = "serif", size =20),          -->
<!-- #         panel.grid.major.x = element_blank(), -->
<!-- #         panel.grid.major.y = element_line( size=.1 ))  -->
<!-- # ``` -->
<!-- #  -->
<!-- # ```{r} -->
<!-- # fmr1_1 = var_part$BSpan %>%  -->
<!-- #   dplyr::filter(gene_name == "FMR1") %>% -->
<!-- #   dplyr::select(-gene_name) %>%  -->
<!-- #   column_to_rownames("EnsemblID") %>%  -->
<!-- #   melt()  %>%  -->
<!-- #   ggplot(aes(x=variable, y = value)) + geom_bar(stat = "identity", fill = "#F16160") + -->
<!-- #   coord_flip() + theme_bw() + ylab("Fraction Variance Explained") + xlab("") + ggtitle("FMR1 - BrainSpan") + -->
<!-- #      theme(axis.text.x = element_text(size = 12, family = "serif", color = "black", angle = 90), -->
<!-- #         axis.text.y = element_text(size = 15, family = "serif", color = "black"),  -->
<!-- #         axis.title = element_text(size=15, family = "serif"),  -->
<!-- #          legend.box.background = element_rect(color = "black"), -->
<!-- #           legend.text = element_text(family = "serif"),  -->
<!-- #           legend.title = element_text( family = "serif"), -->
<!-- #            plot.title = element_text(family = "serif", size =20),          -->
<!-- #         panel.grid.major.x = element_blank(), -->
<!-- #         panel.grid.major.y = element_line( size=.1 ))  -->
<!-- # ``` -->
<!-- #  -->
<!-- # ```{r} -->
<!-- # upf1_2 =var_part$GTEx %>%  -->
<!-- #   dplyr::filter(gene_name == "UPF1") %>% -->
<!-- #   dplyr::select(-gene_name) %>%  -->
<!-- #   column_to_rownames("EnsemblID") %>%  -->
<!-- #   melt()  %>%  -->
<!-- #   ggplot(aes(x=variable, y = value)) + geom_bar(stat = "identity", fill = "#F16160") + -->
<!-- #   coord_flip() + theme_bw() + ylab("Fraction Variance Explained") + xlab("") + ggtitle("UPF1 - GTEx") + -->
<!-- #      theme(axis.text.x = element_text(size = 12, family = "serif", color = "black", angle = 90), -->
<!-- #         axis.text.y = element_text(size = 15, family = "serif", color = "black"),  -->
<!-- #         axis.title = element_text(size=15, family = "serif"),  -->
<!-- #          legend.box.background = element_rect(color = "black"), -->
<!-- #           legend.text = element_text(family = "serif"),  -->
<!-- #           legend.title = element_text( family = "serif"), -->
<!-- #            plot.title = element_text(family = "serif", size =20),          -->
<!-- #         panel.grid.major.x = element_blank(), -->
<!-- #         panel.grid.major.y = element_line( size=.1 ))  -->
<!-- # ``` -->

<!-- # ```{r} -->
<!-- # fmr1_2 = var_part$GTEx %>%  -->
<!-- #   dplyr::filter(gene_name == "FMR1") %>% -->
<!-- #   dplyr::select(-gene_name) %>%  -->
<!-- #   column_to_rownames("EnsemblID") %>%  -->
<!-- #   melt()  %>%  -->
<!-- #   ggplot(aes(x=variable, y = value)) + geom_bar(stat = "identity", fill = "#F16160") + -->
<!-- #   coord_flip() + theme_bw() + ylab("Fraction Variance Explained") + xlab("") + ggtitle("FMR1 - GTEx") + -->
<!-- #      theme(axis.text.x = element_text(size = 12, family = "serif", color = "black", angle = 90), -->
<!-- #         axis.text.y = element_text(size = 15, family = "serif", color = "black"),  -->
<!-- #         axis.title = element_text(size=15, family = "serif"),  -->
<!-- #          legend.box.background = element_rect(color = "black"), -->
<!-- #           legend.text = element_text(family = "serif"),  -->
<!-- #           legend.title = element_text( family = "serif"), -->
<!-- #            plot.title = element_text(family = "serif", size =20),          -->
<!-- #         panel.grid.major.x = element_blank(), -->
<!-- #         panel.grid.major.y = element_line( size=.1 ))  -->
<!-- # ``` -->

<!-- # ```{r} -->
<!-- # ggsave(upf1_1, file= "/home/neuro/Documents/Brain_integrative_transcriptome/BITHub-preprocessing/output/Poster_plots/varPart-upf1-bs.svg", height = 10, width =8, units =  -->
<!-- #          "in") -->
<!-- # ggsave(upf1_2, file= "/home/neuro/Documents/Brain_integrative_transcriptome/BITHub-preprocessing/output/Poster_plots/varPart-upf1-gtex.svg", height = 10, width =8, units =  -->
<!-- #          "in") -->
<!-- # ggsave(fmr1_1, file= "/home/neuro/Documents/Brain_integrative_transcriptome/BITHub-preprocessing/output/Poster_plots/varPart-fmr1-bs.svg", height = 10, width =8, units =  -->
<!-- #          "in") -->
<!-- #  -->
<!-- # ggsave(fmr1_2, file= "/home/neuro/Documents/Brain_integrative_transcriptome/BITHub-preprocessing/output/Poster_plots/varPart-fmr1-gtex.svg", height = 10, width =8, units =  -->
<!-- #          "in") -->
<!-- # ``` -->
